+++ mocks/users_random.json
@@ -8,7 +8,7 @@
   },
   "response": {
     "body": [
-      { "id": 1, "name": "#random:string:length=10&symbol=false#" },
+      { "id": 1, "name": "#random:word:count=5#" },
       { "id": 2, "name":  "#random:name:male=true#", "age": "#random:int:min=10&max=100#", "email": "#random:email#" }
     ]
   }
