+++ src/main/resources/application.properties
@@ -1,3 +1,3 @@
-faker.version=2.1.1
+faker.version=2.1.2
 faker.source=./mocks
 faker.watch=false
