+++ src/main/resources/words.json
@@ -1,0 +1,41 @@
+[
+  "periodic",
+  "twist",
+  "care",
+  "blossom",
+  "hideous",
+  "board",
+  "dry",
+  "hapless",
+  "winter",
+  "hypnotize",
+  "pleasant",
+  "cushion",
+  "pest",
+  "cord",
+  "parcel",
+  "arise",
+  "trousers",
+  "old",
+  "inform",
+  "unknown",
+  "fail",
+  "handsomely",
+  "walk",
+  "lake",
+  "dog",
+  "story",
+  "turkey",
+  "unruly",
+  "moan",
+  "smelly",
+  "expand",
+  "uncovered",
+  "literate",
+  "maniacal",
+  "drug",
+  "frogs",
+  "exuberant",
+  "worm",
+  "grumpy"
+]
